Fix serve-stale backport
When the serve-stale feature was backported to 9.11-S, part of its code
was placed in the wrong location inside query_find().  This flaw causes
an infinite recursion loop to be triggered when serving stale answers is
enabled and no stale answer to a query is available in the cache.

Fix by moving the misplaced code section into its proper location.