address assertion failure by counting tcpalive more carefully
this change causes the tcpalive counter to be incremented whenever a TCP
socket is attached to a client - this includes both client->tcplistener
and client->tcpsock - and decremented whenver *either*
client->tcplistener or client->tcpsock is detached.

in the previous commit, pipelining could cause the tcpalive value to get
out of sync with active TCP clients. now, it always, and only, goes to 0
when the interface is shut down.

I also incidentally put braces around if statements in the code blocks
that I touched.