[WIP] Do not truncate existing files when saving keys
dst_key_tofile() and dst__privstruct_writefile() open their destination
files using fopen() with mode "w", which causes the opened file to be
truncated.  These functions are used by certain DNSSEC utilities which
work on existing files, e.g. dnssec-settime.  If such a utility is used
to modify a key file just as named tries to read it, the latter may
consider the file to be empty and abort or postpone the operation in
progress.  To prevent this, make the aforementioned functions write the
key to a temporary file first (in the same directory as the destination
file) and then move that temporary file to its final location, making it
impossible for the key file to become empty at any point.

This change means that DNSSEC utilities working on existing key files
now require the directory containing the modified key files to be
writable.

TODO:

  - CHANGES note
  - perhaps needs a release note?
  - perhaps needs an ARM update?