Implement digest_rrsig for ZONEMD
Prior to ZONEMD there was never a need to digest a RRSIG as they
where only computed and components compared.  ZONEMD treats RRSIG
as data with the exception of RRSIG(ZONEMD).

Note: the signer name in a RRSIG is case sensitive for DNSSEC order
comparison so compare_rrsig and casecompare_rrsig should be identical.