autosign: use FIPS compatible algorithms and key sizes
Note: the 'bar' zone was not converted.  We need to find a pair
of RSASHA265 keys with similar keyid properties.  The subtest is
currently skipped in fips mode.

Note: the nsec.example zone was not converted as we use it to
test nsec-only to nsec3 conversion failure.  The subtest is skipped
in fips mode.