Modify FIPS mode to depend on runtime detection
RHEL does not enforce FIPS mode at build time. Detection on buildtime
would be unusable for us. Modify it to autodetection mode. If FIPS
related functions are available, test the mode. If FIPS is enabled,
restrict some algorithm and key sizes. If not, continue as usual.

Keep the configure time option, but make it forcing FIPS at build time.
Makes it possible to have different build, which would have FIPS
algorithms restricted even when the crypto library does not require it.