fix the fetches-per-server quota calculation
since commit bad5a523c2e, when the fetches-per-server quota
was increased or decreased, instead of the value being set to
the newly calculated quota, it was set to the *minimum* of
the new quota or 1 - which effectely meant it was always set to 1.
it should instead have been the maximum, to prevent the value from
ever dropping to zero.