Make all tasks to be bound to a thread
Previously, the task could be created unbound or bound to a specific
thread (worker loop).  The unbound tasks would be assigned random thread
every time the isc_task_send() would be called.  Because there's no
logic that would assigned the task to the least busy worker, this just
creates an unpredictability.  Instead of the random assignment, bind all
the previously unbound tasks to worker 0 which must always exists.