rpm: handle upstream and package versions
The upstream version is detected by apkg from tarball name and is
exposed to the jinja2 template as the `{{ version }}` variable.

For special cases, where the upstream version has some dashes in it
(e.g. 9.16.29-S1, 9.19.3-dev), there has to be a separate variable for
the package version. The only difference is that the package version
uses dots instead of dashes in the version string.