Prepare the new "allowed-member-zones" configuration option for catz
Add "allowed-member-zones" to the list of supported options for
configuring a catalog zone.

The option should support a list of names (also wildcard) of the
member zones that are allowed for the catalog zone consumer to accept
from the catalog zone producer. Each element can be negated with the
"!" character.