Process the "allowed-member-zones" configuration option for catz
Make `named` parse and process the "allowed-member-zones" configuration
option for catalog zones.

Implement the list as a reference counted structure to avoid copying
its data into every member zone's instance of the options object.

Introduce a new attribute for `dns_name_t`, so that catz can
differentiate between a normal and negated names. Negated names are
marked in the configuration file by prepending the name with the `!`
character.

When left un-configured, the "allowed-member-zones" option defaults
to allow all member zone names, so this new options should be
backwards-compatible.