Use the "allowed-member-zones" configuration option when processing catz
Implement a functoin to check a given name with the allowed member zones
list and return a boolean value indicating whether the name is allowed
or not.

The negated elements of the list have more weight than the normal
elements.

Also tweak the wording of an existing comment to make it more
understandable.