Prevent unexpected client read callbacks
In the loopmgr branch, there was a change to stop reading from the UDP
socket only after the client read callback was finished to prevent
datagrams being lost between the stop operation and the read operation
in the callback.  The negative side effect was that now the extra
datagram could indeed be received between the read callback and the stop
operation.

Check if the isc_nm_read() was called (sock->recv_read) in the client
read callback and don't call the read callback if nobody is expecting it.