Increase priority of statement definitions in search results
Value 0 currently provides highest "score boost" and makes statement
definitions appear on top of seach results, which seems sensible to me.
With this change search for "max-cache-size" now returns its definition
at the very top.

This change is very obscure because "priority" field in Sphinx API is passed
through several layers down to client-side JavaScript.
In the end the "priority" value stored in search index is used by
performObjectSearch() function in JavaScript, which then **uses the
priority value as an index into Scorer.objPrio array** and adds value
from this array to the current "score". Priority 0 - thus index 0 - has
a code comment "used to be importantResults" and indeed has highest
score out of available values.

Beware: ReadTheDocs uses server-side search and might or might not be
affected by this change.