Don't sign the raw zone
The raw zone is not supposed to be signed.  DNSKEY records in a raw zone
should not trigger zone signing.  The update code needs to be able to
identify when it is working on a raw zone.  Add dns_zone_israw() to
enable it to do this.

(cherry picked from commit 4b287ac02170351ce16f0d60a89b95a7f84ab0f2)