Use thread-local queues for dns_adb LRU updates
Expand the LRU queue with thread-local buffers that are used to store
the name and entries buckets when write lock hasn't been acquired in the
get_{name,entry}bucket.  The write lock is acquired when the space in
the thread-local buffer runs out of space, we are adding a new bucket or
we are in the overmem condition.  The buckets in the thread-local
buffers are then moved at the beginning of the LRU list in slightly
imperfect order as we walk the full buffers in a loop for each thread.

Additionally, don't update the LRU (neither in global LRU nor in the
local LRU buffers) if the bucket was already used after the last LRU
buffer flush.