Make worker->shuttingdown atomic
This record member could be accessed concurrently from within the
context of multiple threads. For example, it is used in
isc__nmsocket_closing(), which can be called on listener sockets from
a context of a thread which does not own the object.

This lead to TSAN reports, this commit should fix that.