Rework the fetch context hash tables and locking
Remove the zone buckets as there was only a single name in each bucket.
Add .lock to fctx and use it when manipulating the fetch context data
and use the bucket lock only when adding/removing the fetch contexts to
the bucket.

Release the bucket (and counters) from the hash tables and memory when the
last fetch context releases the bucket (and counter).