Simplify the name compression hash set
Instead of Robin Hood hashing, use triangular probing.

This removes some complications from the inner loops, at the cost of
possibly needing more loops owing to collisions. However, the hash
sets are generously sized so collisions are less likely to be a
problem than they are for hash sets that try to minimize empty slots.