Spread the implicitly created nodes into all buckets
The implicitly created interior RBT nodes that were created when
splitting the tree were all assigned locknum == 0.  This would make the
bucket 0 overloaded when cleaning the RBT.

Add an argument to dns_rbt_create() to indicate the number of buckets,
and assign the node into the bucket when the rbtnode is created.