Refactor dns_rbt_delete{node,name} to return parent to cleanup
Change the RBTDB <-> RBT integration in a way that RBT indicates to
remove the layer violations where RBTDB looks at .parent, .down and
.data nodes and deciding whether we might want to cleanup the parent.

Instead, make the dns_rbt_delete{node,name} to return the parent that
might be cleaned up to the caller in the extra argument and leave the
decision whether to cleanup to the caller.  This is needed because RBT
doesn't know anything about the reference counting, so it doesn't know
if the empty parent node is or is not in use.  Only the caller knows
that.

As an accompanying change remove the unused 'recurse' argument from the
dns_rbt_delete{node,name}.