Introduce new db type for storing delegations
Introduce a new database type dns_dbtype_delegation, that is going to
be used for storing and looking up delegation records (NS).

This database will be used for the resolver to follow delegations. NS
RRsets will be stored into both the cache db and delegation db. The
child side of the NS RRset will be stored in the cache database, while
the parent side will be stored in the delegation cache.