Use uv_udp_try_send() to reduce latency when sending UDP
Use the synchronous UDP sending when possible as this should reduce the
latency as we are going to avoid polling the network interface on the
send.

Unfortunately, it's not yet possible to call the isc_nm_send() callbacks
synchronously on success, so we still have to allocate uvreq and call
the send callback asynchronously.