Add test cases for MAX_TTL
Add test cases checking that the zone's maximum TTL is increased when
a higher TTL is encountered (when reading the zone, or when accepting
an update).

Reuse template2.db.in for the dynamic-inline.kasp test cases. To do
so, the SOA serial needs to be high enough to be accepted.