Checked array allocation arithmetic with isc_mem_get and friends
Reduce the temptation to calculate allocation sizes unsafely by
changing the `isc_mem_get()` / `isc_mem_put()` family of functions to
allocate arrays by default. That is, they now take two size arguments:
a count of elements and the size of each element, like `calloc()`.
Unlike `calloc()` you must explicitly pass `ISC_MEM_ZERO` if you want
the memory cleared.

The overflow checks are implemented inside the macros in `<isc/mem.h>`
so that the compiler can see that the element size is constant: it
should always be `sizeof(something)`.

There are some new `fx` versions of the functions, for allocating
structures with flexible array members. These function have a third
size argument, `size_t base`, which is the size of the structure, and
`count` and `size` describe the array.